WIP [ci skip] clean sub menu components factored out
