+++ content/src/fieldConfig/config.json
@@ -56,7 +56,7 @@
     "myNJAccountText": "myNewJersey account",
     "addBusinessButton": "Add a Business",
     "navBarUnnamedForeignRemoteSellerWorkerText": "Unnamed Out-of-State Business",
-    "navBarGuestText": "Guest",
+    "navBarGuestText": "Guest Business",
     "registerButton": "Get Started",
     "needsAccountSnackbarBody": "Now that you've filled out your business information, you can save your progress [link your Business.NJ.gov account with myNewJersey](/account-setup?source=guest_snackbar). ",
     "logInButton": "Log in",
