+++ web/src/components/navbar/NavBar.tsx
@@ -1,11 +1,12 @@
-import { NavBarDesktop } from "@/components/navbar/NavBarDesktop";
-import { NavBarLandingDesktop } from "@/components/navbar/NavBarLandingDesktop";
-import { NavBarLogoOnly } from "@/components/navbar/NavBarLogoOnly";
-import { NavBarMobile } from "@/components/navbar/NavBarMobile";
+import { NavBarDesktop } from "@/components/navbar/desktop/NavBarDesktop";
+import { NavBarMobile } from "@/components/navbar/mobile/NavBarMobile";
 import { MediaQueries } from "@/lib/PageSizes";
 import { Task } from "@/lib/types/types";
 import { useMediaQuery } from "@mui/material";
-import { ReactElement, useEffect, useState } from "react";
+import { useRouter } from "next/router";
+import { ROUTES } from "@/lib/domain-logic/routes";
+import { ReactElement, useContext, useEffect, useMemo, useState } from "react";
+import { AuthContext } from "@/contexts/authContext";
 
 type Props = {
   landingPage?: boolean;
@@ -19,6 +20,14 @@ type Props = {
 export const NavBar = (props: Props): ReactElement => {
   const isLargeScreen = useMediaQuery(MediaQueries.desktopAndUp);
   const [scrolled, setScrolled] = useState(false);
+
+  const { state } = useContext(AuthContext);
+  const isAuthenticated = useMemo(() => {
+    return state.isAuthenticated === "TRUE";
+  }, [state.isAuthenticated]);
+
+
+  const router = useRouter();
 
   const handleScroll = (): void => {
     const offset = window.scrollY;
@@ -36,12 +45,24 @@ export const NavBar = (props: Props): ReactElement => {
     };
   }, []);
 
-  if (props.logoOnly) {
-    return <NavBarLogoOnly logoType={props.logoOnly} />;
-  } else if (props.landingPage && isLargeScreen) {
-    return <NavBarLandingDesktop />;
-  } else if (isLargeScreen) {
-    return <NavBarDesktop previousBusinessId={props.previousBusinessId} />;
+  const currentlyOnboarding = (): boolean => {
+    if (!router) {
+      return false;
+    }
+    return router.pathname === ROUTES.onboarding;
+  };
+
+
+  if (isLargeScreen) {
+    return (
+      <NavBarDesktop
+        isLanding={props.landingPage}
+        logoOnlyType={props.logoOnly}
+        previousBusinessId={props.previousBusinessId}
+        currentlyOnboarding={currentlyOnboarding()}
+        isAuthenticated={isAuthenticated}
+      />
+    );
   } else {
     return (
       <>
@@ -52,6 +73,9 @@ export const NavBar = (props: Props): ReactElement => {
           hideMiniRoadmap={props.hideMiniRoadmap}
           isLanding={props.landingPage}
           previousBusinessId={props.previousBusinessId}
+          logoOnlyType={props.logoOnly}
+          currentlyOnboarding={currentlyOnboarding()}
+          isAuthenticated={isAuthenticated}
         />
         <div className={scrolled ? "padding-top-6" : ""} />
       </>
