+++ web/src/components/navbar/desktop/NavBarDesktopHomeLogo.tsx
@@ -1,0 +1,27 @@
+import { NavBarDashboardLink } from "@/components/navbar/NavBarDashboardLink";
+import { NavBarVerticalLine } from "@/components/navbar/NavBarVerticalLine";
+import { NavbarBusinessNjGovLogo } from "@/components/navbar/NavbarBusinessNjGovLogo";
+import { useConfig } from "@/lib/data-hooks/useConfig";
+import { ReactElement } from "react";
+
+
+interface Props {
+  previousBusinessId: string | undefined;
+}
+
+export const NavBarDesktopHomeLogo = (props: Props): ReactElement => {
+
+  const { Config } = useConfig(); // Maybe pass this instead?
+
+  return (
+  <div className="display-flex flex-row flex-align-center">
+    <NavbarBusinessNjGovLogo />
+    <div className="margin-x-105">
+      <NavBarVerticalLine />
+    </div>
+    <NavBarDashboardLink
+      linkText={Config.navigationDefaults.navBarMyAccountText}
+      previousBusinessId={props.previousBusinessId}
+    />
+  </div>);
+}
