+++ web/src/lib/domain-logic/getNavBarBusinessTitle.ts
@@ -1,5 +1,4 @@
 import { getMergedConfig } from "@/contexts/configContext";
-import { IsAuthenticated } from "@/lib/auth/AuthContext";
 import { Business, LookupIndustryById, LookupLegalStructureById } from "@businessnjgovnavigator/shared";
 import {
   isOwningBusiness,
@@ -8,10 +7,10 @@ import {
 
 export const getNavBarBusinessTitle = (
   business: Business | undefined,
-  isAuthenticated: IsAuthenticated
+  isAuthenticated: boolean,
 ): string => {
   const Config = getMergedConfig();
-  if (!business || isAuthenticated === IsAuthenticated.FALSE) {
+  if (!business || !isAuthenticated) {
     return Config.navigationDefaults.navBarGuestText;
   }
