+++ web/src/styles/sections/right-nav.scss
@@ -92,8 +92,7 @@
     margin: 0;
     padding: 0;
     flex: 0 1 auto;
-    padding-left: 0.75rem;
-    padding-right: 0.75rem;
+    margin-right: 0.75rem;
     color: $primary;
     background-color: transparent;
     font-size: 0.81rem;
